requirements.txt
fastapi
sqlmodel
uvicorn
pyyaml